+++ packages/flowr_dart/lib/src/mixin/run_catching.dart
@@ -1,22 +1,18 @@
 import 'dart:async';
 
-mixin RunCatchingMx{
+mixin RunCatchingMx {
   FutureOr<R?> runCatching<R>(
     FutureOr<R> Function() block, {
-    void Function(R data)? onSuccess,
-    void Function(Object e, StackTrace s)? onFailure,
+    FutureOr<R?> Function(R data)? onSuccess,
+    FutureOr<R?> Function(Object e, StackTrace s)? onFailure,
   }) async {
     try {
       final data = block();
-      if (data is Future<R>) {
-        await data.then((e) => onSuccess?.call(e));
-      } else {
-        onSuccess?.call(data);
-      }
-      return data;
+      return data is Future<R>
+          ? await data.then((e) => onSuccess?.call(e))
+          : onSuccess?.call(data);
     } catch (e, s) {
-      onFailure?.call(e, s);
-      return null;
+      return onFailure?.call(e, s);
     }
   }
 }