+++ packages/flowr_dart/CHANGELOG.md
@@ -1,2 +1,6 @@
-## 1.0.0
+## 1.1.0 2025-8-23
+* refactor FlowR::put will return value;
+* fix ::runCatching return value;
+
+## 1.0.0 2025-8-23
 * Split pure dart part form [FlowR](https://github.com/Hu-Wentao/flowr)